Round Icon default, getTracks bug fixed

+++ app/src/main/java/eu/selfhost/riegel/superfit/database/DataBase.kt
@@ -31,8 +31,8 @@ object DataBase {
                             return@exec asMapSequence().map {
                                 Track((it[TrackTable.ID] as Long).toInt(),
                                         it[TrackTable.TrackName] as String? ?: "",
-                                        it[TrackTable.Distance] as Float? ?: 0F,
-                                        it[TrackTable.AverageSpeed] as Float? ?: 0F,
+                                        (it[TrackTable.Distance] as Double? ?: 0.0).toFloat(),
+                                        (it[TrackTable.AverageSpeed] as Double? ?: 0.0).toFloat(),
                                         it[TrackTable.Time] as Long)
                             }.toList().toTypedArray()
                         }
